[feat] BaseTime 엔티티 common/entity/ 경로로 이동. topic, category 엔티티 생성 #52
