+++ backend/src/main/java/com/springboot/domain/auth/service/AuthServiceImpl.java
@@ -8,7 +8,7 @@ import com.springboot.domain.auth.model.SignDto;
 import com.springboot.domain.common.error.exception.BusinessException;
 import com.springboot.domain.common.error.exception.ErrorCode;
 import com.springboot.domain.common.error.exception.InvalidValueException;
-import com.springboot.domain.common.model.ResponseDto;
+import com.springboot.domain.common.model.dto.ResponseDto;
 import com.springboot.domain.common.model.SuccessCode;
 import com.springboot.domain.common.service.ResponseService;
 import com.springboot.domain.member.model.Member;
