+++ backend/src/main/java/com/springboot/domain/common/model/dto/ResponseDto.java
@@ -1,0 +1,24 @@
+package com.springboot.domain.common.model.dto;
+
+import com.springboot.domain.common.model.SuccessCode;
+import java.util.ArrayList;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+@AllArgsConstructor
+public class ResponseDto {
+    private int status;
+    private String message;
+    private String code;
+    private Object data;
+
+    public static ResponseDto commonResponse(SuccessCode code) {
+        return new ResponseDto(code.getStatus(), code.getMessage(), code.getCode(), new ArrayList<>());
+    }
+    public static ResponseDto commonResponse(SuccessCode code, Object body) {
+        return new ResponseDto(code.getStatus(), code.getMessage(), code.getCode(), body);
+    }
+}
