+++ backend/src/main/java/com/springboot/domain/common/service/ResponseService.java
@@ -1,6 +1,6 @@
 package com.springboot.domain.common.service;
 
-import com.springboot.domain.common.model.ResponseDto;
+import com.springboot.domain.common.model.dto.ResponseDto;
 import com.springboot.domain.common.model.SuccessCode;
 import org.springframework.http.ResponseEntity;
