+++ backend/src/main/java/com/springboot/domain/relation/service/RelationServiceImpl.java
@@ -1,14 +1,13 @@
 package com.springboot.domain.relation.service;
 
 import com.springboot.domain.auth.model.UserDetailsImpl;
-import com.springboot.domain.common.model.ResponseDto;
+import com.springboot.domain.common.model.dto.ResponseDto;
 import com.springboot.domain.common.model.SuccessCode;
 import com.springboot.domain.common.service.ResponseService;
 import com.springboot.domain.member.model.Member;
 import com.springboot.domain.member.service.MemberService;
 import com.springboot.domain.relation.model.Relation;
 import com.springboot.domain.relation.repository.RelationRepository;
-import java.util.Optional;
 import javax.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
