+++ backend/src/main/java/com/springboot/domain/topic/model/entity/Topic.java
@@ -1,0 +1,33 @@
+package com.springboot.domain.topic.model.entity;
+
+import com.springboot.domain.common.model.entity.BaseTime;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+@Entity
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+@ToString
+public class Topic extends BaseTime {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(columnDefinition = "TEXT", nullable = false)
+    private String name;
+
+    @Column(nullable = false)
+    private int count;
+
+}
